+++ CHANGELOG.md
@@ -164,6 +164,10 @@ All notable changes to this project will be documented in this file.
 -   #2125 : Fix missing type check for argument of known type in a function with arguments whose type can be one of several types.
 -   #2042 : Fix missing index in loop unravelling for loop of length 1.
 -   #2093 : Fix scoping issue preventing class methods from sharing a name with locals in another class method.
+-   #1814 : Fix class method visitation order to correctly access the global scope from methods.
+-   #1668 : Fix handling of `is not None` check to ensure it is always checked before accessing the variable.
+-   #802 : Add if blocks in Python output to ensure support for implementations that differ for different types.
+-   Fix casting of arrays in Python translation.
 
 ### Changed